do not delete locked SMS from the MCE

+++ android/app/src/com/android/bluetooth/map/BluetoothMapContentObserver.java
@@ -2769,6 +2769,32 @@ public class BluetoothMapContentObserver {
         return res;
     }
 
+    private boolean checkSmsIsLocked(long handle) {
+        boolean locked = false;
+        Uri uri = ContentUris.withAppendedId(Sms.CONTENT_URI, handle);
+        Cursor c = null;
+        try {
+            c = mResolver.query(uri, new String[] { Sms.LOCKED }, null, null, null);
+            if (c != null) {
+                if (c.moveToFirst()) {
+                    if (c.getInt(0)==1) {
+                        locked = true;
+                        Log.w(TAG, "checkSmsIsLocked,locked = " + locked);
+                    }
+                }
+            }
+        } catch (SecurityException e) {
+            Log.e(TAG, "Catching SecurityException : " + e);
+        } catch (CursorWindowAllocationException e) {
+            Log.e(TAG, "Catching CursorWindowAllocationException : " + e);
+        } finally {
+            if (c != null) {
+                c.close();
+            }
+        }
+        return locked;
+    }
+
     @VisibleForTesting
     boolean deleteMessageSms(long handle) {
         boolean res = false;
@@ -2890,6 +2916,9 @@ public class BluetoothMapContentObserver {
         } else {
             if (statusValue == BluetoothMapAppParams.STATUS_VALUE_YES) {
                 if (type == TYPE.SMS_GSM || type == TYPE.SMS_CDMA) {
+                    if (checkSmsIsLocked(handle)) {
+                        return false;
+                    }
                     res = deleteMessageSms(handle);
                 } else if (type == TYPE.MMS) {
                     res = deleteMessageMms(handle);
